Change multiple sizing declarations.
I gave the body tag of the page a margin in percentages instead of pixels.
This fixed the alignment issue I had with the ul in the header. Added flex-wrap
to the "cards" tag so no the page will reorient itself on page resize, rather
than coming out of center and creating side scroll.